+++ swarm/src/connection.rs
@@ -32,9 +32,10 @@ pub use error::{
 pub use listeners::{ListenersEvent, ListenersStream};
 pub use pool::{ConnectionCounters, ConnectionLimits};
 pub use pool::{EstablishedConnection, PendingConnection};
-pub use substream::{Close, Substream, SubstreamEndpoint};
+pub use substream::{Close, SubstreamEndpoint};
 
 use crate::handler::ConnectionHandler;
+use crate::IntoConnectionHandler;
 use handler_wrapper::HandlerWrapper;
 use libp2p_core::connection::ConnectedPoint;
 use libp2p_core::multiaddr::Multiaddr;
@@ -94,11 +95,20 @@ where
     /// Builds a new `Connection` from the given substream multiplexer
     /// and connection handler.
     pub fn new(
+        peer_id: PeerId,
+        endpoint: ConnectedPoint,
         muxer: StreamMuxerBox,
-        handler: THandler,
+        handler: impl IntoConnectionHandler<Handler = THandler>,
         substream_upgrade_protocol_override: Option<upgrade::Version>,
+        max_negotiating_inbound_streams: usize,
     ) -> Self {
-        let wrapped_handler = HandlerWrapper::new(handler, substream_upgrade_protocol_override);
+        let wrapped_handler = HandlerWrapper::new(
+            peer_id,
+            endpoint,
+            handler,
+            substream_upgrade_protocol_override,
+            max_negotiating_inbound_streams,
+        );
         Connection {
             muxing: Muxing::new(muxer),
             handler: wrapped_handler,