+++ protocols/kad/CHANGELOG.md
@@ -1,3 +1,15 @@
+# 0.38.0 [unreleased]
+
+- Update to `libp2p-core` `v0.34.0`.
+
+- Update to `libp2p-swarm` `v0.37.0`.
+
+# 0.37.1
+
+- Limit # of inbound streams to 32. [See PR 2699].
+
+[PR 2699]: https://github.com/libp2p/rust-libp2p/pull/2699
+
 # 0.37.0
 
 - Update to `libp2p-core` `v0.33.0`.