+++ protocols/relay/CHANGELOG.md
@@ -1,3 +1,20 @@
+# 0.10.0 [unreleased]
+
+- Update to `libp2p-core` `v0.34.0`.
+
+- Update to `libp2p-swarm` `v0.37.0`.
+
+# 0.9.1
+
+- Respond to at most one incoming reservation request. Deny <= 8 incoming
+  circuit requests with one per peer. And deny new circuits before accepting new
+  circuits. See [PR 2698].
+
+- Expose explicits errors via `UpgradeError` instead of generic `io::Error`. See
+  [PR 2698].
+
+[PR 2698]: https://github.com/libp2p/rust-libp2p/pull/2698/
+
 # 0.9.0
 
 - Update to `libp2p-core` `v0.33.0`.